reflection-probe-manager move to 3d module (#15359)
* move to 3d module

* export probetype

* fix error
