+++ cocos/3d/reflection-probe/index.ts
@@ -23,3 +23,5 @@
 */
 
 export { ReflectionProbe } from './reflection-probe-component';
+export { ReflectionProbeManager } from './reflection-probe-manager';
+export { ReflectionProbeType } from './reflection-probe-enum';
