+++ cocos/3d/reflection-probe/reflection-probe-component.ts
@@ -28,7 +28,7 @@ import { CCBoolean, CCObject, Color, Enum, Vec3 } from '../../core';
 import { TextureCube } from '../../asset/assets';
 import { scene } from '../../render-scene';
 import { CAMERA_DEFAULT_MASK } from '../../rendering/define';
-import { ReflectionProbeManager } from '../../rendering/reflection-probe-manager';
+import { ReflectionProbeManager } from './reflection-probe-manager';
 import { Component } from '../../scene-graph/component';
 import { Layers } from '../../scene-graph/layers';
 import { Camera } from '../../misc/camera-component';
