+++ cocos/3d/reflection-probe/reflection-probe-manager.ts
@@ -24,16 +24,16 @@
 */
 
 import { EDITOR } from 'internal:constants';
-import { MeshRenderer } from '../3d/framework/mesh-renderer';
-import { ReflectionProbeType } from '../3d/framework/reflection-probe-enum';
-import { ImageAsset, Texture2D } from '../asset/assets';
-import { PixelFormat } from '../asset/assets/asset-enum';
-import { Vec3, geometry, cclegacy } from '../core';
-import { AABB } from '../core/geometry';
-import { Texture } from '../gfx';
-import { Camera, Model } from '../render-scene/scene';
-import { ProbeType, ReflectionProbe } from '../render-scene/scene/reflection-probe';
-import { Layers } from '../scene-graph/layers';
+import { MeshRenderer } from '../framework/mesh-renderer';
+import { ReflectionProbeType } from './reflection-probe-enum';
+import { ImageAsset, Texture2D } from '../../asset/assets';
+import { PixelFormat } from '../../asset/assets/asset-enum';
+import { Vec3, geometry, cclegacy } from '../../core';
+import { AABB } from '../../core/geometry';
+import { Texture } from '../../gfx';
+import { Camera, Model } from '../../render-scene/scene';
+import { ProbeType, ReflectionProbe } from '../../render-scene/scene/reflection-probe';
+import { Layers } from '../../scene-graph/layers';
 
 const REFLECTION_PROBE_DEFAULT_MASK = Layers.makeMaskExclude([Layers.BitMask.UI_2D, Layers.BitMask.UI_3D, Layers.BitMask.GIZMOS, Layers.BitMask.EDITOR,
     Layers.BitMask.SCENE_GIZMO, Layers.BitMask.PROFILER, Layers.Enum.IGNORE_RAYCAST]);
